Bastion - Initial work creating engine to hold new UI assets.
This commit introduces the idea of an inline engine dedicated to
UI components. The first component moved is the Systems and Nutupane
implementations. In time this engine will serve as the place where
new UI goes or re-factorings get moved to. Ultimately, the entire UI
portion should be contained within this engine.