Bastion: Adding HTML5 boilerplate based base template.
Adds the current HTML5 boilerplate as the base template and adjusts
the styling to accomodate this base layout change. This change is intended
to be as close to current HTML5 best practices as possible and provide
the initial full-page width for nutupane.

Adds a hook from the system model that the consumer modules in the glue
layer can hook in to and add information to the json structure from.