fixes #4874 - content view filter: display validation error to user
This commit is to ensure that validation errors when
creating a new filter are displayed to the user.  The
approach used is similar/same as what is done with
creating a new content view, in that the validation
will be inlined in the form, such that the user may
correct the error and resubmit.