Fixes #4482: Adds manual bootstrapping and modules specification for the Angular app.
Given the boostrap process can only occur once, declaring all our module
requirements in an array and manually bootstrapping provides the ability
to dynamically add modules for inclusion. This in turn will allow secondary
or plugin modules to declare themselves and be included within the greater
application. This can be achived by the following:

BASTION_MODULES.push('myModuleName')

This should not place any requirements on when the JavaScript containing the
module declaration is loaded as long as it is prior to the DOM content being
loaded.