fixes #6624 / BZ 1105617 - content host package actions - disable buttons when busy
This commit contains a minor change to disable the ability to initiate a new action,
until the current action has been accepted (202) by the server.  The most common
use case for this is if the user were to double-click the 'Perform', 'Update All'
or '(X)' items in the UI.

Without this change, the user might observe an error like the following in the log:
  ForemanTasks::Lock::LockConflict: Required lock is already taken by other running tasks.

In spite of the above error, the user could later go back and re-execute a new request.