fixes #5096 - fail gracefully on pulp repo creation failure
The goal is to run the action creation in the plan phase
and error out if its not valid.  A similar approach is taken
with Content Host Creation.  However in that case, content hosts
are not created in other situations.  Repositories are created during
publish/promote of content views as well and in that situation we do not
want to perform pulp creation in the plan phase.  As a result I have created
a new action CreateImmediate that does the creation in the plan phase.
It attempts to use as much from Create as possible (there is a lot of code there)