Fixes #7294: replace <th> translate filters with directives.
We were using translate filters in our table headers because
of a clash between the translate directive and other directives.
This commit updates the text in the <th>s to be wrapped in a
span that includes the translate directive instead.  Directives
are more performant than filters so this should speed up the
application ever so slightly.

http://projects.theforeman.org/issues/7294