fixes #7309 / BZ 1093483 - Repo sync status - update UI cross-links and status based on dynflow task
The Sync Status page was recently updated to sync repositories
using dynflow tasks.  With this, all syncs whether they are
initiated by Sync Status, Product or Repository pages should
be executed through dynflow.

That said, the API (which is used for showing sync status on
the Products, Repositories, Content Views and Content View
Filter pages) is currently reporting the status using
pulp; however, that does not provide a true indication of
whether or not the sync process has completed successfully.

This commit will address 2 things:

1. Update the API to return the 'last sync' information using the
   dynflow task (vs pulp). This information is then used for displaying
   sync time, status...etc.
2. Update the cross-linking for individual repositories on the various
   pages to go directly to the dynflow task versus sending the user to the
   sync status page.  This will allow the user to see
   any errors that may have occured without having to search for
   the repo once landing on Sync Status and then having to click
   again to get to the dynflow task.