Fixes #4546: Align table headers properly when scrollbar exists with sticky header.
This introduces the idea of adjusting the table width based on the scroll
width of the alch container scroll to prevent data from hiding behind the scroll bar.
Further, a blank table head is added to the table header to push the header
columns back to the left and properly align the grid.