Fixes #9729: speed up nutupane by removing JS BZ 1161630.
There was some non-performant JS that was causing nutupane
to be really slow when there were lots of rows in a table,
namely triggering resize on each row inserted into the table.
This commit replaces the slow JS with an approach that is more
in performant and more in line with Angular JS best practices.

http://projects.theforeman.org/issues/9729
https://bugzilla.redhat.com/show_bug.cgi?id=1161630
(cherry picked from commit 18718747079605837eeebf1f5ba2fef3d3787fb1)