Fixes #10973 - Add the translations from downstream to upstream.
This takes a cleaned version of the downstream stirngs and adds them to upstream.
To accomplish this, I first extracted the strings and then ran a command which pulled
teh strings from downstream into upstream.

I would like to see this merged, and then the strings sent to an external translation
service. In the past this was zanata. That file was out of date, so I deleted it. Ideally,
these would go into a Transifex server to be consistent with foreman.