Fixes #10515: check if scroll is needed after results are loaded.
After replacing our own infinite scroll with ngInfiniteScroll
we weren't checking to see if scroll was needed after nutupane
results were loaded.  This could cause the display of only one
set of results on larger screens.  This commit forces ngInfiniteScroll
to recheck if a scroll is needed after results are loaded.

http://projects.theforeman.org/issues/10515